DevOps/CI CD/лекция_нетологии/example with Django/web-chat/requirements-dev.txt
-r requirements.txt
flake8
wemake-python-styleguide
pytest
pytest-django
